Rename ArcTo file
