+++ src/index.ts
@@ -14,5 +14,5 @@ export * from './components/circle.tsx';
 export * from './components/line.ts';
 export * from './components/path.ts';
 export * from './components/point.ts';
-export * from './components/arc.ts';
+export * from './components/arc-to.ts';
 export * from './create-context.ts';
